# src/main/resources/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Api Documentation",
    "version": "1.0",
    "title": "Api Documentation",
    "termsOfService": "urn:tos",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "paths": {
...
    "/account/{userId}/proposal": {
      "get": {
        "tags": ["proposal-controller"],
        "summary": "getAllProposalByUserId",
        "operationId": "getAllProposalByUserIdUsingGET",
        "consumes": ["application/json"],
        "produces": ["*/*"],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "userId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {"$ref": "#/definitions/Page«Proposal»"}
          },
          "401": {"description": "Unauthorized"},
          "403": {"description": "Forbidden"},
          "404": {"description": "Not Found"}
        }
      },
      "post": {
        "tags": ["proposal-controller"],
        "summary": "createProposal",
        "operationId": "createProposalUsingPOST",
        "consumes": ["application/json"],
        "produces": ["*/*"],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "userId",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "proposal",
            "description": "proposal",
            "required": true,
            "schema": {"$ref": "#/definitions/Proposal"}
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {"type": "object"}
          },
          "201": {"description": "Created"},
          "401": {"description": "Unauthorized"},
          "403": {"description": "Forbidden"},
          "404": {"description": "Not Found"}
        }
      }
    },
...
    "/city/{cityId}/proposal": {
      "get": {
        "tags": ["proposal-controller"],
        "summary": "getAllProposalByFromCityId",
        "operationId": "getAllProposalByFromCityIdUsingGET",
        "consumes": ["application/json"],
        "produces": ["*/*"],
        "parameters": [
          {
            "name": "cityId",
            "in": "path",
            "description": "cityId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {"$ref": "#/definitions/Page«Proposal»"}
          },
          "401": {"description": "Unauthorized"},
          "403": {"description": "Forbidden"},
          "404": {"description": "Not Found"}
        }
      }
    },
...
  },
  "definitions": {
...
    "City": {
      "type": "object",
      "properties": {
        "country": {"type": "string"},
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {"type": "string"}
      }
    },
...
    "Page«Proposal»": {
      "type": "object",
      "properties": {
        "content": {
          "type": "array",
          "items": {"$ref": "#/definitions/Proposal"}
        },
        "first": {"type": "boolean"},
        "last": {"type": "boolean"},
        "number": {
          "type": "integer",
          "format": "int32"
        },
        "numberOfElements": {
          "type": "integer",
          "format": "int32"
        },
        "size": {
          "type": "integer",
          "format": "int32"
        },
        "sort": {"$ref": "#/definitions/Sort"},
        "totalElements": {
          "type": "integer",
          "format": "int64"
        },
        "totalPages": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "Proposal": {
      "type": "object",
      "properties": {
        "delivery": {"type": "string"},
        "description": {"type": "string"},
        "fromCityId": {
          "type": "integer",
          "format": "int64"
        },
        "fromcity": {"$ref": "#/definitions/City"},
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "idUser": {
          "type": "integer",
          "format": "int64"
        },
        "name": {"type": "string"},
        "price": {"type": "string"},
        "shop": {"$ref": "#/definitions/Shop"},
        "status": {"type": "boolean"},
        "toCityId": {
          "type": "integer",
          "format": "int64"
        },
        "tocity": {"$ref": "#/definitions/City"},
        "type": {"type": "string"},
        "user": {"$ref": "#/definitions/User"}
      }
    },
    "Shop": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "logo": {"type": "string"},
        "name": {"type": "string"}
      }
    },
    "Sort": {"type": "object"},
    "User": {
      "type": "object",
      "properties": {
        "city": {"$ref": "#/definitions/City"},
        "fullName": {"type": "string"},
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "password": {"type": "string"},
        "role": {"type": "string"},
        "username": {"type": "string"}
      }
    },
...
  }
}